% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatToml.R
\name{formatTOML}
\alias{formatTOML}
\title{Format data into a TOML string}
\usage{
formatTOML(data)
}
\arguments{
\item{data}{[list] list of values}
}
\value{
A TOML-formatted string
}
\description{
The function \code{formatTOML} formats given data (a list) into TOML-formatted string.
}
\details{
This implementation uses the \sQuote{tomlplusplus} library by Mark Gillar (see
\url{https://github.com/marzer/tomlplusplus}) which is TOML 1.0 compliant.
}
