% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatToml.R
\name{formatTOML}
\alias{formatTOML}
\title{Format data into a TOML string}
\usage{
formatTOML(data, width = 0L)
}
\arguments{
\item{data}{[list] list of values to encode as TOML}

\item{width}{[integer] maximum width of a line in TOML output,
affects array inlining}
}
\value{
A TOML-formatted string
}
\description{
The function \code{formatTOML} formats given data (a list) into
TOML-formatted string.
}
\details{
This implementation uses the \sQuote{toml11} library by Toru Niina
(see \url{https://github.com/ToruNiina/toml11}), which is
TOML 1.0 compliant.
}
